Added MembersWithUserIds spec to clean up leaderboard
