--- src/DevBetterWeb.Core/Specs/BookByIdWithMembersSpec .cs
@@ -1,14 +1,0 @@
-﻿using Ardalis.Specification;
-using DevBetterWeb.Core.Entities;
-
-namespace DevBetterWeb.Core.Specs
-{
-    public class BookByIdWithMembersSpec : Specification<Book>
-    {
-        public BookByIdWithMembersSpec(int bookId)
-        {
-            Query.Where(book => book.Id == bookId);
-            Query.Include(book => book.MembersWhoHaveRead);
-        }
-    }
-}
